Create UserSourceController
